Accounted for Glenn Renfro's feedback
Glenn caught a number of minor errors and one big problem. The big problem was that I tried to change the heading level down one (there should be only one first-level heading in the book). I can't do it in all the files without changing a lot of stuff in other projects, so that will have to wait for a separate effort. Consequently, I reduced the headings by one (back to their original depth).